Emit backtraces for all calls, including slot callbacks
Increase the verbosity further in the server logs, and provide the full
backtrace (or as much as we can get) for each AM call. This isn't sent
to the client, so it requires log_min_messages=debug2 or higher.

Additionally, add a TRACE to each tts_* callback.

Static functions are not included, which can lead to sparse backtraces,
but with unstripped binaries, a command like

    objdump --syms /path/to/postgres | sort -n

can help track down the missing frames by address.