Change slot synchronization technique
Rather than rely on a single slot synchronization method to get it right
every time, explicitly copy slot fields that have already been shown to
change in practice, and prohibit changes to any other fields.

This will make it more obvious when an assumption that has been made is
wrong, and it provides living documentation of which fields change in
practice (though work is still needed to document _why_).