XXX: add (broken) test for CREATE INDEX
This is the end of the line for this implementation strategy, for now.
The problem is that, for all the effort put into ensuring that the
objects here are compatible with the heap AM, there are a couple of
functions that hardcode a check on the type to try to ensure that you've
passed them the "correct" type of object.

In this case, we try to start a heap scan -- which goes though
heap_getnext(), which doesn't go through the table AM -- and we get

    ERROR:  only heap AM is supported

as the test failures here show.

We might be able to just set the toast AM to the heap AM to try to keep
going further, but I think at that point we're reaching diminishing
returns for the method in general.